test: adapt CplexMipDemonstratorTest to general DemonstratorTest
